bits_bug

+++ include/lib/stdint.h
@@ -11,6 +11,6 @@ typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
 typedef unsigned long uint64_t;
 
-typedef unsigned int size_t;
+typedef unsigned long size_t;
 
 #endif
